Move Prover future into its own module
